# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FordSpec AI — API",
    "description": "API REST do Desafio 01 (Inteligência Competitiva Automotiva). Recebe Marca + Modelo + Versão + atributos e devolve uma ficha técnica padronizada a partir de uma base curada, com regras determinísticas (zero alucinação). Campos inexistentes saem como 'ND'.",
    "contact": {
      "name": "Equipe FordSpec — FIAP"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health check",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/attributes": {
      "get": {
        "tags": [
          "catalog & vehicles"
        ],
        "summary": "Lista o dicionário de atributos (13 categorias)",
        "description": "catalog-service — retorna os ~244 atributos padronizados.",
        "operationId": "list_attributes_v1_attributes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttributeOut"
                  },
                  "type": "array",
                  "title": "Response List Attributes V1 Attributes Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/vehicles": {
      "get": {
        "tags": [
          "catalog & vehicles"
        ],
        "summary": "Lista versões disponíveis na base curada",
        "description": "vehicle-service — versões curadas para a marca/modelo informados.",
        "operationId": "list_vehicles_v1_vehicles_get",
        "parameters": [
          {
            "name": "brand",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "examples": [
                "Ford"
              ],
              "title": "Brand"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "Ranger"
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleOut"
                  },
                  "title": "Response List Vehicles V1 Vehicles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/specs": {
      "post": {
        "tags": [
          "specs"
        ],
        "summary": "Gera a ficha técnica padronizada",
        "description": "spec-service — recebe Marca + Modelo + Versão + atributos (livre)\ne devolve a ficha no formato padronizado.",
        "operationId": "create_spec_v1_specs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecSheet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/specs/{request_id}": {
      "get": {
        "tags": [
          "specs"
        ],
        "summary": "Recupera os metadados de uma ficha gerada",
        "operationId": "get_spec_v1_specs__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/specs/export": {
      "get": {
        "tags": [
          "specs"
        ],
        "summary": "Exporta a ficha no formato BASE (CSV)",
        "description": "export-service — gera CSV no layout do data sheet (Categoria; Atributo; Valor).",
        "operationId": "export_spec_v1_specs_export_get",
        "parameters": [
          {
            "name": "brand",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Ford",
              "title": "Brand"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Ranger",
              "title": "Model"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AttributeOut": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "type": "object",
        "required": [
          "category",
          "name"
        ],
        "title": "AttributeOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SpecItem": {
        "properties": {
          "attribute": {
            "type": "string",
            "title": "Attribute"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "attribute",
          "category",
          "value",
          "status"
        ],
        "title": "SpecItem"
      },
      "SpecRequestIn": {
        "properties": {
          "brand": {
            "type": "string",
            "title": "Brand",
            "examples": [
              "Ford"
            ]
          },
          "model": {
            "type": "string",
            "title": "Model",
            "examples": [
              "Ranger"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "examples": [
              "Limited 3.0L V6 26MY"
            ]
          },
          "attributes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes",
            "description": "Lista livre de atributos. Se vazia, retorna a ficha completa.",
            "examples": [
              [
                "Potência",
                "Torque",
                "Tração"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "brand",
          "model",
          "version"
        ],
        "title": "SpecRequestIn"
      },
      "SpecSheet": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "vehicle": {
            "additionalProperties": true,
            "type": "object",
            "title": "Vehicle"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "total_attributes": {
            "type": "integer",
            "title": "Total Attributes"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SpecItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle",
          "generated_at",
          "total_attributes",
          "items"
        ],
        "title": "SpecSheet"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehicleOut": {
        "properties": {
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "version": {
            "type": "string",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "brand",
          "model",
          "version"
        ],
        "title": "VehicleOut"
      }
    }
  }
}